working Iintop_imm and added polling print
Iintop_imm is called right after Ialloc to update young_ptr stored
in rax. Now, rax is treated as a live register with valid value.
Inserting a poll before this instruction means that if the
poll causes a minor heap empyting, then the rax value is pointing
to an invalid object.

Possible solutions -

1. Emit the information that the value in the register is of type
Addr until the header is inited. This would ensure that the poll
is not inserted until atleast the header is inited
2. Infer at the poll stage to not insert any polls from the alloc
to header write. This seems a little problematic, because in the case
the process of initing a block a changes, the polling pass would break

Exploring various options